backend/requirements.txt
fastapi
uvicorn[standard]
torch          
diffusers
transformers
accelerate
git+https://github.com/openai/point-e.git
open3d
Pillow
tqdm
numpy
scipy